Fix AVL insert. Add Tree Clone. Adjust test layouts.
AVL insert was failing to handle the case where the root node of the rotations needed to be updated. Cleaned up and simplified some of the code to properly handle this.
BST Delete implemented with tests.
Update all existing tests structures to include a test case name and use t.Run in the for test := range tests loops.
Added Clone as a primary method on Tree to eliminate the need for crazy reflection and generic unwrapping in order to clone a given Tree.
Cleaned up comments in various files.
Split dump binary tree ASCII functions to separate file.
Added more tests for Binary Trees Equal/Equivalent.
Added some helpers to test files to clean up repetition.